Add cancel button to TaskForm
A Cancel button has been added to the TaskForm component of the Task page. On clicking this Cancel button, users will be navigated back to the previous page. This improvement provides users with an alternative way to stop creating a task without needing to refresh the page or use browser navigation buttons.