Refactor UI Code and Update Task handling
This commit does two main things. First, it simplifies the return value of our Button component rendering function in 'Button.tsx' to remove an unnecessary fragment. Secondly, in the 'EditTaskPage.tsx' file, it modifies the initial state value for the 'id' property of taskToUpdate, and adds a conditional render when no id exists to better handle non-existent tasks.