Refactored SWR usage in frontend code
This commit modifies how SWR is used across frontend components. Previously, SWRConfig was used at a global level in App.tsx to define a fetcher function for fetching data with SWR. This has been removed, and the fetcher function is now defined locally wherever useSWR is called, which is currently in taskData.tsx. In addition, modifications have been made in TaskDetailsPage.tsx to call the SWR mutate function directly from the swr import instead of from the useSWRConfig hook, and error handling has been introduced for both the SWR mutate function and for a delete task Axios call.